+++ src/javascript/utils/filesystem.js
@@ -1,4 +1,5 @@
 import requirePath from '../utils/require'
+import StaticConfig from 'StaticConfig'
 
 function openFolder(){
     return new Promise((resolve, reject) => {
@@ -9,6 +10,14 @@ function openFolder(){
             })
             .then(result => {
                 if (result.canceled) return;
+                
+                if(!StaticConfig.data.log.filter((a)=>a.directory ==result.filePaths[0])[0]){
+                    StaticConfig.data.log.push({
+                        directory:result.filePaths[0]
+                    })
+                }
+               
+                StaticConfig.triggerChange()
                 resolve(result.filePaths[0])
             })
             .catch(err => {