+++ src/javascript/utils/static.config.js
@@ -1,15 +1,14 @@
 import {puffin} from '@mkenzo_8/puffin'
 import {initConfiguration , getConfiguration} from './configurator'
 
-initConfiguration()
 const cachedConfiguration = getConfiguration()
 
-let editorConf = Object.assign({},cachedConfiguration.editor)
+let editorConf = Object.assign({},cachedConfiguration.config)
 
 const StaticConfig = new puffin.state(editorConf)
 
 StaticConfig.changed(function(data){
-    cachedConfiguration.store.set('editor',data)
+    cachedConfiguration.store.set('config',StaticConfig.data)
 })
 
 export default StaticConfig